server/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.9.5"
python-socketio = "^5.11.2"
pydantic = "^2.7.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.2"
pyright = "^1.1.360"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
extend-select = ["I"]
